upgrade bintrees
The bintrees@1.0.1 package doesn't include full license text in the distribution. Transitively, tdigest is affected, because its dependency isn't properly licensed. bintrees@1.0.2 solves the issue. In some enterprise environments, missing full license text in the distribution complicates usage of the package as a dependency, especially given the MIT license itself mentions "The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software."